Ported compression working on host, separated compression and decompression to different files
