--- snappy/host-compress/.gitignore
@@ -1,8 +1,0 @@
-*.o
-*.so
-scmd
-sgverify
-verify
-bench
-fuzzer
-ftracer
