--- snappy/host-compress/util.h
@@ -1,9 +1,0 @@
-#include <stdio.h>
-
-void *xmalloc(size_t size);
-char *basen(char *s);
-int compare(char *a, char *b, size_t size);
-
-
-#define err(x) perror(x), exit(1)
-
